[pkg/stanza] Move operatortest package into operator directory
This package was located within the operator/helper directory.
Moving this to be more in line with other test packages in the repo.
